requirements-base.txt
asdf>=2.15.0
asdf-standard==1.0.3
asdf-transform-schemas==0.3.0
asdf-unit-schemas==0.1.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==22.12.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.5
django-ckeditor==6.7.0
django-distill==3.1.3
django-grappelli===3.0.8
filelock==3.9.0
Flask==2.3.3
Flask-Markdown==0.3
Flask-WTF==1.1.1
gunicorn==21.2.0
h5py==3.9.0
identify==2.5.15
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy_loader==0.3
Markdown==3.4.4
MarkupSafe==2.1.3
mypy-extensions==0.4.3
networkx==3.1
nodeenv==1.7.0
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.0
Pillow==10.0.0
platformdirs==2.6.2
pre-commit==3.0.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.2
semantic-version==2.10.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.4
tifffile==2023.8.12
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.17.1
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
